restructure to remove unrelated items from button folder
