--- packages/guideline-blocks-settings/src/components/BlockItemWrapper/Toolbar/ToolbarAttachments.tsx
@@ -1,33 +1,0 @@
-/* (c) Copyright Frontify Ltd., all rights reserved. */
-
-import { Attachments } from '../../Attachments';
-import { useAttachmentsContext } from '../../../hooks';
-
-import { AttachmentsToolbarButton } from './ToolbarButton/AttachmentsToolbarButton';
-import { useToolbarFlyoutState } from './hooks/useToolbarFlyoutState';
-import { useMemoizedId } from '@frontify/fondue';
-
-export const ToolbarAttachments = () => {
-    const id = useMemoizedId();
-
-    const { appBridge, attachments, onAttachmentsAdd, onAttachmentDelete, onAttachmentReplace, onAttachmentsSorted } =
-        useAttachmentsContext();
-
-    const { isOpen, onOpenChange } = useToolbarFlyoutState(id);
-
-    return (
-        <Attachments
-            onUpload={onAttachmentsAdd}
-            onDelete={onAttachmentDelete}
-            onReplaceWithBrowse={onAttachmentReplace}
-            onReplaceWithUpload={onAttachmentReplace}
-            onSorted={onAttachmentsSorted}
-            onBrowse={onAttachmentsAdd}
-            items={attachments}
-            appBridge={appBridge}
-            triggerComponent={AttachmentsToolbarButton}
-            isOpen={isOpen}
-            onOpenChange={onOpenChange}
-        />
-    );
-};
