+++ packages/guideline-blocks-settings/src/components/BlockItemWrapper/Toolbar/ToolbarButton/ToolbarButton.tsx
@@ -2,7 +2,7 @@
 
 import { useIsDragPreview } from '../context/DragPreviewContext';
 import { BaseToolbarButton } from './BaseToolbarButton';
-import { ToolbarButtonTooltip } from './ToolbarButtonTooltip';
+import { ToolbarButtonTooltip } from '../ToolbarButtonTooltip';
 
 export type ToolbarButtonProps = { icon: JSX.Element; tooltip?: string; onClick: () => void };
