+++ packages/guideline-blocks-settings/src/components/BlockItemWrapper/Toolbar/ToolbarButton/index.ts
@@ -4,4 +4,4 @@ export * from './DragHandleToolbarButton';
 export * from './ToolbarButton';
 export * from './MenuToolbarButton';
 export * from './FlyoutToolbarButton';
-export * from './FlyoutMenu';
+export * from './AttachmentsToolbarButton';
